fix error messages
