+++ test/canisters/fs_tests/src/fs_tests_backend/src/main.rs
@@ -1,5 +1,4 @@
-mod lib;
-
+mod canister;
 fn main() {
-    lib::basic_fs_test();
+    canister::basic_fs_test();
 }
